+++ src/render.js
@@ -125,12 +125,15 @@ export function renderSidebar (content) {
 
   if (content) {
     html = markdown(content)
+    // find url tag
+    html = html.match(/<ul[^>]*>([\s\S]+)<\/ul>/g)[0]
   } else if (OPTIONS.sidebar) {
     html = tpl.tree(OPTIONS.sidebar, '<ul>')
   } else {
     html = tpl.tree(genTree(toc, OPTIONS.maxLevel), '<ul>')
   }
 
+  html = (OPTIONS.name ? `<h1><a href="${OPTIONS.nameLink}">${OPTIONS.name}</a></h1>` : '') + html
   renderTo('aside.sidebar', html)
   const target = event.activeLink('aside.sidebar', true)
   if (target) renderSubSidebar(target)
@@ -154,18 +157,18 @@ export function renderCover (content) {
     return
   }
   renderCover.dom.classList.add('show')
-  if (renderCover.rendered) return
+  if (renderCover.rendered) return event.sticky()
 
   // render cover
   let html = markdown(content)
-  const match = html.trim().match('<p><img[^s]+src="(.*?)"[^a]+alt="(.*?)"></p>$')
+  const match = html.trim().match('<p><img[^s]+src="(.*?)"[^a]+alt="(.*?)">([^<]*?)</p>$')
 
   // render background
   if (match) {
     const coverEl = document.querySelector('section.cover')
 
     if (match[2] === 'color') {
-      coverEl.style.background = match[1]
+      coverEl.style.background = match[1] + (match[3] || '')
     } else {
       coverEl.classList.add('has-mask')
       coverEl.style.backgroundImage = `url(${match[1]})`